[#22] Move primitive <-> conservative transform into FluidPhysics.
