+++ src/stanshock/physics/fluid_base.py
@@ -6,6 +6,8 @@ from dataclasses import dataclass
 import cantera as ct
 import numpy as np
 
+from stanshock.system.base import Array
+
 
 @dataclass
 class FluidState:
@@ -178,3 +180,45 @@ class FluidPhysics(ABC):
     @abstractmethod
     def get_source_terms(self, state: FluidState):
         """Compute reaction source terms corresponding to transported scalars."""
+
+    def primitive_to_conservative(self, state: FluidState):
+        """Transform primitive variables into vector of conservatives."""
+        return np.concatenate(
+            (
+                state.density[..., None],
+                (state.density * state.velocity)[..., None],
+                (
+                    state.pressure / (state.gamma - 1.0)
+                    + 0.5 * state.density * state.velocity**2
+                )[..., None],
+                state.density[..., None] * state.composition,
+            ),
+            axis=-1,
+        )
+
+    def conservative_to_primitive(self, state_array: Array, gamma: Array) -> FluidState:
+        """Transform conservative variables into primitives."""
+        r = state_array[..., 0]
+        ru = state_array[..., 1]
+        E = state_array[..., 2]
+        rY = state_array[..., 3:]
+
+        u = ru / r
+        p = (gamma - 1.0) * (E - 0.5 * r * u**2.0)
+        Y = rY / r[..., None]
+
+        # Bound
+        Y[Y > 1.0] = 1.0
+        Y[Y < 0.0] = 0.0
+
+        # Scale
+        if self.normalize_scalars:
+            Y = Y / np.sum(Y, axis=-1, keepdims=True)
+
+        return FluidState(
+            shape=r.shape,
+            density=r,
+            velocity=u,
+            pressure=p,
+            composition=Y,
+        )
